Update syllabus
The reading list has been overhauled to reflect what worked and what
didn't in Fall 2014.  The course description has been punched up too.
The structure of assignments is different for Spring 2015 due to higher
enrollment---doing a final exam instead of a paper.

Also a technical change: the Markdown syllabus is no longer
hard-wrapped.  Of course this messes with tracking; sorry.